discovery-service/requirements.txt
# Core deps — mirrors the main backend's discovery subset
python-jobspy>=0.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
pandas>=2.0.0
playwright>=1.40.0
httpx>=0.25.0

# Turso/libSQL support (required when DATABASE_URL=libsql://...)
libsql-experimental>=0.0.8
